% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fred.totalunemployment.R
\docType{data}
\name{fred.totalunemployment}
\alias{fred.totalunemployment}
\title{fred.totalunemployment contains the historical monthly data on Total
unemployed, plus all marginally attached workers plus total employed
part time for economic reasons from FRED from 1/1/1994 to 12/1/2014.}
\description{
fred.totalunemployment contains the historical monthly data on Total
unemployed, plus all marginally attached workers plus total employed
part time for economic reasons from FRED from 1/1/1994 to 12/1/2014.
}
\author{
Federal Reserve Bank of St. Louis
}
\references{
\url{http://research.stlouisfed.org/fred2/series/U6RATENSA#}
}
\keyword{data}

