% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PPNRValidation-package.R
\docType{package}
\name{PPNRValidation-package}
\alias{PPNRValidation}
\alias{PPNRValidation-package}
\title{Simplified PPNR validation}
\description{
The PPNRValidation package allows for easy validation of PPNR CCAR models.
}
\author{
Omar Azhar, Anton Bossenbroek, David Karapetyan, Prabakar Rajasekaran
}

