% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.StationarityTest.R
\name{print.StationarityTest}
\alias{print.StationarityTest}
\title{Prints the results of the stationarity tests.}
\usage{
\method{print}{StationarityTest}(object, format = NULL,
  return.statistics = FALSE, ...)
}
\arguments{
\item{object}{The StationarityTest result.}

\item{format}{The format the table in a format given by kable.}

\item{return.statistics}{Indicate whether thedata.frame that contains all the
statistics shoudl be returned.}

\item{...}{Additional parameters passed on to kable.}
}
\value{
return a data.frame with test name, series name, test statistic and
p-value for each test performed.
}
\description{
This function prints the outcome of the various stationarity tests.
}
\examples{
data(fred.totalunemployment)
 result <- stationarity.test(fred.totalunemployment)
 print(result)

 print(result, format = "latex")
}
\seealso{
knitr::kable
}

