% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/stationarity.test.R
\name{stationarity.test}
\alias{stationarity.test}
\title{Standard tests to verify whether a time series is stationary.}
\usage{
stationarity.test(time.series)
}
\arguments{
\item{time.series}{a time series that should be tested.}
}
\value{
object of class StationarityTest
}
\description{
This function plots the ACF and PACF plots for a time series and performs
the Ljung-Box, Augmented Dickey-Fuller (ADF) and
Kwiatkowski-Phillips-Schmidt-Shin (KPSS) tests.
}
\examples{
data(fred.totalunemployment)
 result <- stationarity.test(fred.totalunemployment)
}

